Add grpc port and split logic

+++ internal/entity/departament.go
@@ -3,11 +3,11 @@ package entity
 import "time"
 
 type Department struct {
-	ID          uint      `json:"id,omitempty"`
+	ID          uint64    `json:"id,omitempty"`
 	Name        string    `json:"name"`
 	Description string    `json:"description"`
-	ParentID    *uint     `json:"parent_id"`
-	HeadID      *uint     `json:"head_id"`
+	ParentID    *uint64   `json:"parent_id"`
+	HeadID      *uint64   `json:"head_id"`
 	CreatedAt   time.Time `json:"created_at,omitempty"`
 	UpdatedAt   time.Time `json:"updated_at,omitempty"`
 }
